Init ChartViewModel only when needed
